Refactor: Push down schedulePush methods to TablesRepository
- Moved schedulePush(Account) from AbstractRepository to TablesRepository
- Moved schedulePush(Account, SyncStatusReporter) from AbstractRepository to TablesRepository
- Only TablesRepository uses these methods (10 usages), not AccountRepository or SearchRepository
- Changed treeSyncAdapter visibility to protected in AbstractRepository
- Location: AbstractRepository.java (removed lines 38-48), TablesRepository.java (added lines 63-73)
- No functional changes

Refactoring Type: Push-Down Method (Set II - Refactoring #4)
